Narrow starting_spot_at_angle per operation, taking it exact against the game (#279) (#283)
`starting_spot_at_angle` evaluated in plain f64 with no narrowing anywhere. The
noise machine evaluates it one operation at a time in f32, so the port was a
different computation - and because the function is SHARED between Fulgora and
Vulcanus, it could not be fixed inside a Fulgora-only change.

## The headline

**The direct oracle test goes 88 of 152 to 152 of 152.** Every captured case the
game produced, bit-exact. The comment on that assertion used to explain the 88
away as "the same known port gap the elevation chain carries"; it was not a gap
in the chain, it was this expression. Both ports now assert the same 152 from
the same fixture, with worst residual exactly 0.

Five narrowings are needed and no subset is enough: per-operation, an f32 `pi`,
f32 `sin`/`cos`, and then f32 radius/distance and f32 angle at the CALL SITES.
An f32 `pi` alone helps the vault cone and HURTS the main one until the angle is
narrowed too, which reads like the two cones wanting opposite fixes - they call
the same function, so that reading has to be wrong. It is an artifact of scoring
one change at a time.

## Measured, all re-run on this tree

- Both Fulgora cones 83/101 and 85/101 -> **101/101 at residual exactly 0**,
  with `fulgora_vault_pyramids` 85 -> 101 and `vault_pyramids_and_start`
  77 -> 101 behind them. **13 frozen counts up, 1 down.**
- Terrain PNG **34,977 -> 34,788** differing pixels of 1,048,576. #273 moved
  this by one pixel the wrong way; this moves it 189 the right way, because the
  cones feed the `mix_*` chain the image is made of.
- Scrap footprint: the one game scrap pixel outside it is gone, `outside`
  1 -> **0**.
- Tile argmax UNCHANGED at 4,915 of 5,057, as #279 predicted.
- WASM and TypeScript renders stay byte-identical, so tier 2 and tier 3 both
  hold across the change.

## Vulcanus: a large improvement that a BOUND called a regression

Exact f32 matches out of 1085: `startingTungsten` 614 -> 1082, `startingCoal`
611 -> 974, `startingCalcite` 547 -> 969, `startingSulfur` 618 -> 1049.

The only thing that got worse is calcite's single worst residual, 2.2888e-5 ->
3.0518e-5, tripping a 3e-5 bound. That outlier is at `(-2332.9, 2333.7)`, where
the field's own value is **-133.94** - one f32 ULP there is 1.53e-5, so the bound
is a TWO-ULP bound at that magnitude, and exactly 2 of 1085 positions exceed it.
That is the far-from-origin f32 coordinate floor the other Vulcanus specs
document.

Those four assertions are now frozen EXACT COUNTS with the residual kept
underneath. **The bound was not widened; it was replaced by something stronger,
and that is proven rather than claimed**: un-narrowing the calcite radius drops
the count 969 -> 669 while the residual bound passes unchanged. A bound cannot
tell 547 exact from 969 exact.

The Vulcanus call-site audit #279 asks for is done - five resource sites, three
spawn sites, the three spawn angles. It took calcite 669 -> 969 and did not move
those 2 deep-field points.

## The one count that dropped

`fulgoran_rock_probability` 80 -> 79, recorded at the assertion rather than
smoothed, the same way #273 recorded its two. Same class: a still-inexact deep
composite where one position crossed a rounding boundary the unlucky way. It is
not evidence against the change - the expression it descends from is now exact
at all 152 captured cases.

## The ABI fixture

`bearingTrig` narrows every step now, so the four trig words in
`test/fixtures/wasm-request.v1.json` moved. The LAYOUT did not, so ABI_VERSION
stays 1. Per that fixture's own instruction those four were recomputed in Python
from the definition and only THEN compared against the encoder, which agreed on
all four - so it stayed an independent check rather than becoming a
regeneration. Every other byte is untouched.

## What this does NOT do

The 12 candidates #279 lists are still unapplied, and its prediction about them
is not confirmed: it expected `moats`, `vaultSpots` and `spotsPrebanding` to
reach 101/101 once the cones moved, and measured they reach 69, 69 and 98. They
improved; they did not close.

Gate: `pnpm run verify` rc=0, 231 files / 2061 tests passed, 4 skipped.

Closes #279

+++ src/noise/expressions/fulgoraShared.ts
@@ -141,7 +141,7 @@ export function makeFulgoraShared(ctx: FulgoraCtx): FulgoraShared {
 
   // Both cones are placed by map_seed alone, so the starting island sits at a
   // seed-determined bearing from the origin. The vault sits opposite it (+180).
-  const angle = seed0 / 360;
+  const angle = f32(seed0 / 360);
 
   // startingCone is max(0, A, B) over TWO discs: a wide one offset a little way
   // out, and a tight one at distance 1 whose distortion is damped to a quarter.
@@ -152,8 +152,8 @@ export function makeFulgoraShared(ctx: FulgoraCtx): FulgoraShared {
     const dy = wobbleY(x, y);
     const wide = startingSpotAtAngle({
       angle,
-      distance: grid / 30,
-      radius: grid / 1.8,
+      distance: f32(grid / 30),
+      radius: f32(grid / 1.8),
       xDistortion: 1 * dx,
       yDistortion: 1 * dy,
       xFromStart: x,
@@ -162,7 +162,7 @@ export function makeFulgoraShared(ctx: FulgoraCtx): FulgoraShared {
     const tight = startingSpotAtAngle({
       angle,
       distance: 1,
-      radius: grid / 4,
+      radius: f32(grid / 4),
       xDistortion: 0.25 * dx,
       yDistortion: 0.25 * dy,
       xFromStart: x,
@@ -175,9 +175,9 @@ export function makeFulgoraShared(ctx: FulgoraCtx): FulgoraShared {
     Math.max(
       0,
       startingSpotAtAngle({
-        angle: angle + 180,
-        distance: grid / 1.8,
-        radius: grid / 1.8,
+        angle: f32(angle + 180),
+        distance: f32(grid / 1.8),
+        radius: f32(grid / 1.8),
         xDistortion: 1 * wobbleX(x, y),
         yDistortion: 1 * wobbleY(x, y),
         xFromStart: x,
